WIP: refactoring for CV2 db integration
